+++ ppc-amigaos/scripts/libs/libpng16.sh
@@ -2,11 +2,7 @@
 # 
 
 echo -e "${CCPINK}${CCBOLD}\n---> Install libpng16 ${CCEND}";
-	curl -fsSL "https://github.com/salass00/png16_lib/releases/download/V53.3/png16_lib-53.3.lha" -o libpng16.lha && \
-		lha -xfq2 libpng16.lha && \
-		cp -r ./png16_lib/SDK/local/* ${SDK_PATH}/local/ && \
-		ln -s ${SDK_PATH}/local/newlib/lib/libpng16.a \
-			${SDK_PATH}/local/newlib/lib/libpng.a && \
-		ln -s ${SDK_PATH}/local/newlib/lib/pkgconfig/libpng16.pc \
-			${SDK_PATH}/local/newlib/lib/pkgconfig/libpng.pc && \
+	curl -fsSL "https://os4depot.net/share/development/library/graphics/libpng.lha" -o /tmp/libpng.lha && \
+		lha -xfq2 libpng.lha && \
+		cp -r ./SDK/local/* ${SDK_PATH}/local/ && \
 		rm -rf /tmp/*;